refactor(util): rename util -> utilities-url
